fix(campus): keep node-ical out of the client bundle + webpack
The events module broke the build two ways:
- IntegrationsTab (a client component) imported readEventFeeds from
  lib/events.ts, which top-level-imported node-ical — pulling it
  (and `node:crypto`) into the client bundle: UnhandledSchemeError.
- node-ical doesn't survive webpack bundling on the server either
  (`g.BigInt is not a function` at page-data collection).

Fix:
- Split lib/events.ts (pure, client-safe: readEventFeeds /
  formatEventWhen / type) from lib/events-server.ts (the node-ical
  ICS parsing, server-only). The public page imports each from the
  right place.
- next.config.mjs: serverExternalPackages: ["node-ical"] so it loads
  at runtime instead of being bundled.

Verified with a full `next build`.
