+++ apps/campus/lib/events.ts
@@ -1,12 +1,10 @@
-import { expandRecurringEvent, sync } from "node-ical";
-
 /**
- * Campus events — pulled from ICS calendar feeds.
+ * Campus events — client-safe helpers and types.
  *
- * Feed URLs are stored per-campus in `sceneData.eventFeeds`; the
- * public home page fetches + parses them server-side (Next caches
- * the fetch) and shows what's upcoming. ICS is the reliable spine —
- * Google Calendar / Outlook OAuth and Facebook are deferred.
+ * This module is import-safe from client components (it has no Node
+ * dependencies). The actual ICS fetching + parsing — which pulls in
+ * `node-ical` and Node built-ins — lives in `events-server.ts` and
+ * must only be imported from server code.
  */
 export interface CampusEvent {
   id: string;
@@ -19,12 +17,6 @@ export interface CampusEvent {
   location?: string;
 }
 
-/** How far ahead to surface events, and how many to show. */
-const HORIZON_DAYS = 60;
-const MAX_EVENTS = 12;
-/** Per-feed fetch timeout — a slow feed must not hang the page. */
-const FEED_TIMEOUT_MS = 8000;
-
 /** Read a campus's ICS feed URLs from its `sceneData`. */
 export function readEventFeeds(sceneData: unknown): string[] {
   const raw = (sceneData as { eventFeeds?: unknown } | null | undefined)
@@ -33,65 +25,6 @@ export function readEventFeeds(sceneData: unknown): string[] {
   return raw.filter(
     (u): u is string => typeof u === "string" && u.length > 0,
   );
-}
-
-/** ICS string values can be plain or `{ params, val }` — flatten them. */
-function text(value: unknown): string {
-  if (typeof value === "string") return value;
-  if (value && typeof value === "object" && "val" in value) {
-    const v = (value as { val?: unknown }).val;
-    return typeof v === "string" ? v : "";
-  }
-  return "";
-}
-
-/**
- * Fetch + parse the given ICS feeds and return upcoming events,
- * soonest first. Recurring events are expanded; a bad or slow feed
- * is skipped so the rest still render.
- */
-export async function fetchCampusEvents(
-  feedUrls: string[],
-): Promise<CampusEvent[]> {
-  const now = new Date();
-  const horizon = new Date(now.getTime() + HORIZON_DAYS * 86_400_000);
-  const events: CampusEvent[] = [];
-
-  for (const url of feedUrls) {
-    try {
-      const res = await fetch(url, {
-        next: { revalidate: 1800 },
-        signal: AbortSignal.timeout(FEED_TIMEOUT_MS),
-      });
-      if (!res.ok) continue;
-      const parsed = sync.parseICS(await res.text());
-
-      for (const key of Object.keys(parsed)) {
-        const component = parsed[key];
-        if (!component || component.type !== "VEVENT") continue;
-        for (const instance of expandRecurringEvent(component, {
-          from: now,
-          to: horizon,
-        })) {
-          const start = new Date(instance.start);
-          events.push({
-            id: `${key}::${start.toISOString()}`,
-            title: text(instance.summary) || "Event",
-            start: start.toISOString(),
-            end: new Date(instance.end).toISOString(),
-            allDay: instance.isFullDay,
-            location: text(instance.event.location) || undefined,
-          });
-        }
-      }
-    } catch {
-      // Skip an unreachable / malformed feed — others still render.
-    }
-  }
-
-  return events
-    .sort((a, b) => Date.parse(a.start) - Date.parse(b.start))
-    .slice(0, MAX_EVENTS);
 }
 
 /** Format an event's start for display. */
