+++ apps/campus/next.config.mjs
@@ -8,6 +8,10 @@ const nextConfig = {
     "@klorad/engine-mapbox",
     "@klorad/engine-three",
   ],
+  // `node-ical` (ICS event parsing) is a Node library that doesn't
+  // survive webpack bundling — load it at runtime instead of bundling
+  // it into the server output.
+  serverExternalPackages: ["node-ical"],
   webpack(config) {
     config.resolve.alias = {
       ...config.resolve.alias,
